+++ project/package.json
@@ -65,7 +65,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
-    "@types/faker": "^5.5.8",
+    "@types/faker": "5.5.8",
     "@types/jest": "^29.1.2",
     "@types/leaflet": "^1.9.0",
     "@types/lodash": "^4.14.186",
@@ -76,7 +76,7 @@
     "axios-mock-adapter": "^1.21.2",
     "eslint-config-htmlacademy": "^8.0.0",
     "faker": "^5.5.3",
-    "leaflet": "^1.9.2",
+    "leaflet": "1.9.2",
     "typescript": "^4.8.4"
   }
 }
