+++ project/src/mocks/offers.ts
@@ -32,8 +32,6 @@ export const offers: OfferType[] = [
     isFavorite: true,
     isPremium: true,
     location: {
-      /*latitude: 52.3909553943508,
-      longitude: 4.85309666406198,*/
       latitude: 48.8534100,
       longitude: 2.3488000,
       zoom: 8
